refactor: data storage for pod into its own class
